# project/public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sos/dashboard/active-emergencies": {
      "get": {
        "description": "Get active emergencies (filtered by hospital if HOSPITAL role)",
        "operationId": "SosController_getActiveEmergencies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active emergencies retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Staff role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active emergencies for dashboard",
        "tags": [
          "SOS / Emergency Requests"
        ]
      }
    },
...
    "/rescue/dashboard/stats": {
      "get": {
        "description": "Get overall statistics for the rescue dashboard",
        "operationId": "DashboardController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - ADMIN or EMERGENCY_CENTER role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get dashboard statistics",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/rescue/dashboard/active-emergencies": {
      "get": {
        "description": "Get list of all active emergency cases",
        "operationId": "DashboardController_getActiveEmergencies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active emergencies retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active emergencies",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/rescue/dashboard/team-locations": {
      "get": {
        "description": "Get current locations of all rescue teams",
        "operationId": "DashboardController_getTeamLocations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Team locations retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get team locations",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/rescue/dashboard/hospital-capacities": {
      "get": {
        "description": "Get current bed capacity information for all hospitals",
        "operationId": "DashboardController_getHospitalCapacities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Hospital capacities retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get hospital capacities",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/rescue/dashboard/assign-case": {
      "post": {
        "description": "Assign an emergency case to a rescue team or hospital",
        "operationId": "DashboardController_assignCase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Case assigned successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Case or team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign emergency case",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/rescue/dashboard/cancel-case": {
      "post": {
        "description": "Cancel an active emergency case",
        "operationId": "DashboardController_cancelCase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Case cancelled successfully"
          },
          "404": {
            "description": "Case not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel emergency case",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "Emergency Project API",
    "description": "\n      ## Emergency Response Management System API\n      \n      This API provides comprehensive endpoints for managing emergency requests, hospitals, rescue teams, and related services.\n      \n      ### Features:\n      - **Authentication**: OAuth 2.0 (Google, Facebook, Apple) and Email/Password\n      - **Emergency Requests**: Create and manage SOS emergency requests\n      - **Hospital Management**: Hospital capacity, emergency acceptance\n      - **Rescue Teams**: Team management and assignment\n      - **Dashboard**: Real-time statistics and monitoring\n      - **Notifications**: Real-time push notifications\n      - **Reports**: Generate and download reports\n      - **User Settings**: Manage user preferences and settings\n      \n      ### Authentication:\n      Most endpoints require JWT Bearer token authentication. Use the /auth endpoints to obtain tokens.\n    ",
    "version": "1.0",
    "contact": {
      "name": "Emergency Project Team",
      "url": "https://emergency-project.com",
      "email": "support@emergency-project.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.emergency-project.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "AssignCaseDto": {
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string"
          },
          "assignedToId": {
            "type": "string"
          }
        },
        "required": [
          "caseId",
          "assignedToId"
        ]
      },
      "CancelCaseDto": {
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string"
          }
        },
        "required": [
          "caseId"
        ]
      },
...
    }
  }
}